refactor: clean up imports and remove unused constants in actor-related files
